requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
schedule>=1.2.0
lxml>=4.9.0
PyQt6>=6.5.0
cryptography>=41.0.0
